+++ app/fetcher/src/logging.rs
@@ -6,8 +6,8 @@ pub(crate) const TARGET: &str = "erfiume_fetcher";
 pub(crate) struct Logger {
     chat_id: Option<i64>,
     station: Option<String>,
-    threshold: Option<f32>,
-    value: Option<f32>,
+    threshold: Option<f64>,
+    value: Option<f64>,
     error_text: Option<String>,
 }
 
@@ -26,12 +26,12 @@ impl Logger {
         self
     }
 
-    pub(crate) fn threshold(mut self, threshold: f32) -> Self {
+    pub(crate) fn threshold(mut self, threshold: f64) -> Self {
         self.threshold = Some(threshold);
         self
     }
 
-    pub(crate) fn value(mut self, value: f32) -> Self {
+    pub(crate) fn value(mut self, value: f64) -> Self {
         self.value = Some(value);
         self
     }
